# swagger.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthController_getMe","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/credentials":{"post":{"operationId":"CredentialsController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateCredentialRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialResponseDto"}}}}},"tags":["credentials"]},"get":{"operationId":"CredentialsController_list","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CredentialResponseDto"}}}}}},"tags":["credentials"]}},"/credentials/{id}":{"get":{"operationId":"CredentialsController_get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialResponseDto"}}}}},"tags":["credentials"]}},"/uploads/credentials/{filename}":{"get":{"operationId":"UploadsController_serveImage","parameters":[{"name":"filename","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Credentials API","description":"API para registrar y generar credenciales militares","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"api-key":{"type":"apiKey","in":"header","name":"x-api-key"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"admin@example.com"},"password":{"type":"string","example":"password123"}},"required":["email","password"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"password123"},"personId":{"type":"string","example":"person_id_optional"}},"required":["email","password"]},"CreateCredentialRequestDto":{"type":"object","properties":{"fullName":{"type":"string","example":"Juan Perez"},"rank":{"type":"string","example":"Teniente"},"identityNumber":{"type":"string","example":"123456789"},"typeIdentity":{"type":"string","example":"CC"},"unit":{"type":"string","example":"Batallon 1"},"birthDate":{"type":"string","example":"1990-01-01"},"institutionalEmail":{"type":"string","example":"juan.perez@mil.edu"},"credentialTypeCode":{"type":"string","example":"MILITARY"},"phone":{"type":"string","example":"3001234567"},"image":{"type":"string","format":"binary"}},"required":["fullName","identityNumber","typeIdentity","birthDate","institutionalEmail","credentialTypeCode","image"]},"CredentialResponseDto":{"type":"object","properties":{"id":{"type":"string"},"fullName":{"type":"string"},"rank":{"type":"string"},"identityNumber":{"type":"string"},"unit":{"type":"string"},"birthDate":{"type":"string","format":"date-time"},"institutionalEmail":{"type":"string"},"imagePath":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["id","fullName","rank","identityNumber","unit","birthDate","institutionalEmail","imagePath","createdAt","updatedAt"]}}}}